# GitMiner/Documentation/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /gitminer/projects/{id}:
    get:
      tags:
      - Project
      - projects
      - get
      summary: Retrieve one project
      description: Get one project of the database
      operationId: findOne_1
      parameters:
      - name: id
        in: path
        description: Id of the project to find
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not found
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    put:
      tags:
      - Project
      - projects
      - put
      summary: Update a project
      description: Update a project with all the data
      operationId: update_1
      parameters:
      - name: id
        in: path
        description: Id of the project to be updated
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      responses:
        "400":
          description: Bad Request
        "204":
          description: Project updated
        "404":
          description: Not found
    delete:
      tags:
      - Project
      - project
      - delete
      summary: Delete a project
      description: Delete a project with all the data
      operationId: delete_1
      parameters:
      - name: id
        in: path
        description: Id of the project to be deleted
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
        "204":
          description: Project deleted
        "404":
          description: Not found
...
  /gitminer/projects:
    get:
      tags:
      - Project
      - projects
      - get
      summary: Retrieve all Projects
      description: Get all projects in the database
      operationId: findAll
      parameters:
      - name: title
        in: query
        required: false
        schema:
          type: string
      - name: complexLast
        in: query
        required: false
        schema:
          type: boolean
      - name: page
        in: query
        description: page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: order
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: number of elements per page
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Projects found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "404":
          description: Not found
    post:
      tags:
      - Project
      - post
      - project
      summary: Create a project
      description: Create a project with all the data
      operationId: project
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      responses:
        "400":
          description: Bad Request
        "201":
          description: Project created and returned
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Project'
...
components:
  schemas:
    User:
      required:
      - username
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        web_url:
          type: string
    Comment:
      required:
      - body
      - created_at
      type: object
      properties:
        id:
          type: string
        body:
          type: string
        author:
          $ref: '#/components/schemas/User'
        created_at:
          type: string
        updated_at:
          type: string
    Commit:
      required:
      - author_name
      - authored_date
      - committed_date
      - committer_name
      - web_url
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        author_name:
          type: string
        author_email:
          type: string
        authored_date:
          type: string
        committer_name:
          type: string
        committer_email:
          type: string
        committed_date:
          type: string
        web_url:
          type: string
    Issue:
      type: object
      properties:
        id:
          type: string
        ref_id:
          type: string
        title:
          type: string
        description:
          type: string
        state:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        closed_at:
          type: string
        labels:
          type: array
          items:
            type: string
        author:
          $ref: '#/components/schemas/User'
        assignee:
          $ref: '#/components/schemas/User'
        upvotes:
          type: integer
          format: int32
        downvotes:
          type: integer
          format: int32
        web_url:
          type: string
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    Project:
      required:
      - name
      - web_url
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        web_url:
          type: string
        commits:
          type: array
          items:
            $ref: '#/components/schemas/Commit'
        issues:
          type: array
          items:
            $ref: '#/components/schemas/Issue'